add hashids

+++ mars.go
@@ -2,17 +2,19 @@ package mars
 
 import (
 	"errors"
+	"github/ferdiunal/venus"
 
 	"gorm.io/gorm"
 )
 
 type Mars struct {
-	db *gorm.DB
+	db    *gorm.DB
+	venus venus.VenusInterface
 }
 
 type MarsInterface interface {
 	CreateToken(userId uint64, name string, abilities []string) *MarsToken
-	CheckToken(userId uint64, tokenId uint64, token string) (*PersonalAccessToken, error)
+	CheckToken(tokenId uint64, token string) (*PersonalAccessToken, error)
 	RevokeToken(userId uint64, tokenId uint64, token string) error
 }
 
@@ -20,12 +22,12 @@ func (m *Mars) CreateToken(userId uint64, name string, abilities []string) *Mars
 	model := NewPersonalAccessToken(userId, name, abilities)
 	m.db.Create(model)
 
-	return model.GetResult()
+	return model.GetResult(m.venus)
 }
 
-func (m *Mars) CheckToken(userId uint64, tokenId uint64, token string) (*PersonalAccessToken, error) {
+func (m *Mars) CheckToken(tokenId uint64, token string) (*PersonalAccessToken, error) {
 	model := &PersonalAccessToken{}
-	m.db.Where("id = ? and userId = ?", tokenId, userId).Find(model)
+	m.db.Where("id = ? and userId = ?", tokenId).Find(model)
 
 	if ok := model.HashedToken() == token; ok {
 		return model, nil
@@ -35,10 +37,14 @@ func (m *Mars) CheckToken(userId uint64, tokenId uint64, token string) (*Persona
 }
 
 func (m *Mars) RevokeToken(userId uint64, tokenId uint64, token string) error {
-	model, err := m.CheckToken(userId, tokenId, token)
+	model, err := m.CheckToken(tokenId, token)
 
 	if err != nil {
 		return err
+	}
+
+	if model.UserId != userId {
+		return errors.New("token not matched user id") // Burasını değiştir ileride
 	}
 
 	m.db.Delete(model)
@@ -46,8 +52,10 @@ func (m *Mars) RevokeToken(userId uint64, tokenId uint64, token string) error {
 	return nil
 }
 
-func NewMars(db *gorm.DB) MarsInterface {
+func NewMars(db *gorm.DB, salt string, len int) MarsInterface {
+	hashids := venus.New(salt, len)
 	return &Mars{
-		db: db,
+		db:    db,
+		venus: hashids,
 	}
 }
